+++ src/main/java/com/epam/spring/testingapp/controller/QuestionController.java
@@ -1,8 +1,10 @@
 package com.epam.spring.testingapp.controller;
 
+import com.epam.spring.testingapp.dto.AnswerDto;
 import com.epam.spring.testingapp.dto.QuestionDto;
 import com.epam.spring.testingapp.dto.TestDto;
 import com.epam.spring.testingapp.dto.group.OnUpdate;
+import com.epam.spring.testingapp.exception.NotFoundException;
 import com.epam.spring.testingapp.service.QuestionService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -13,6 +15,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import javax.validation.constraints.Min;
 import java.util.List;
+import java.util.Objects;
 
 @Slf4j
 @RestController
@@ -31,7 +34,13 @@ public class QuestionController {
     @GetMapping("/question/{questionId}")
     public QuestionDto find(@PathVariable @Min(1) int questionId) {
         log.info("find({})", questionId);
-        return questionService.find(questionId);
+
+        QuestionDto questionDto = questionService.find(questionId);
+        if(Objects.isNull(questionDto)){
+            throw new NotFoundException("Question with id %s not found".formatted(questionId));
+        }
+
+        return questionDto;
     }
 
     @PostMapping("{testId}/question")
@@ -44,12 +53,20 @@ public class QuestionController {
     @PutMapping("/question/{questionId}")
     public QuestionDto update(@RequestBody @Validated(OnUpdate.class) QuestionDto questionDto, @PathVariable @Min(1) int questionId) {
         log.info("update({}, {})", questionDto, questionId);
+
+        if(Objects.isNull(questionService.find(questionId))){
+            throw new NotFoundException("Question with id %s not found".formatted(questionId));
+        }
         return questionService.update(questionDto, questionId);
     }
 
     @DeleteMapping("/question/{questionId}")
     public void delete(@PathVariable @Min(1) int questionId) {
         log.info("delete({})", questionId);
+
+        if(Objects.isNull(questionService.find(questionId))){
+            throw new NotFoundException("Question with id %s not found".formatted(questionId));
+        }
         questionService.delete(questionId);
     }
 }
